+++ packages/validator-ajv8/test/precompiledValidator.test.ts
@@ -31,11 +31,30 @@ describe('AJV8PrecompiledValidator', () => {
   describe('default options', () => {
     // Use the AJV8PrecompiledValidator
     let validator: AJV8PrecompiledValidator;
-    let resolvedRootSchema: RJSFSchema;
     beforeAll(() => {
       validator = new AJV8PrecompiledValidator(validateFns, rootSchema);
-      // Since the root schema is retrieved in core before calling the validator emulate that here
-      resolvedRootSchema = retrieveSchema(validator, rootSchema, rootSchema);
+    });
+    describe('validator.ensureSameRootSchema', () => {
+      it('using rootSchema returns true', () => {
+        expect(validator.ensureSameRootSchema(rootSchema)).toBe(true);
+      });
+      it('using resolved rootSchema returns true', () => {
+        const resolvedRootSchema = retrieveSchema(validator, rootSchema, rootSchema);
+        expect(validator.ensureSameRootSchema(resolvedRootSchema)).toBe(true);
+      });
+      it('using a different schema throws', () => {
+        const schema: RJSFSchema = {
+          type: 'object',
+          properties: {
+            name: { type: 'string' },
+          },
+        };
+        expect(() => validator.ensureSameRootSchema(schema)).toThrowError(
+          new Error(
+            'The schema associated with the precompiled validator differs from the rootSchema provided for validation'
+          )
+        );
+      });
     });
     describe('validator.isValid()', () => {
       it('should return true if the data is valid against the schema', () => {
@@ -126,14 +145,16 @@ describe('AJV8PrecompiledValidator', () => {
           },
         };
         expect(() => validator.validateFormData({}, schema)).toThrowError(
-          new Error('The schema associated with the precompiled schema differs from the schema provided for validation')
+          new Error(
+            'The schema associated with the precompiled validator differs from the rootSchema provided for validation'
+          )
         );
       });
       describe('No custom validate function, single value of correct type generates no errors', () => {
         let errors: RJSFValidationError[];
 
         beforeAll(() => {
-          const result = validator.validateFormData({ foo: '42' }, resolvedRootSchema);
+          const result = validator.validateFormData({ foo: '42' }, rootSchema);
           errors = result.errors;
         });
 
@@ -146,7 +167,7 @@ describe('AJV8PrecompiledValidator', () => {
         let errorSchema: ErrorSchema;
 
         beforeAll(() => {
-          const result = validator.validateFormData({ foo: 42 }, resolvedRootSchema);
+          const result = validator.validateFormData({ foo: 42 }, rootSchema);
           errors = result.errors;
           errorSchema = result.errorSchema;
         });
@@ -163,7 +184,7 @@ describe('AJV8PrecompiledValidator', () => {
       describe('Validating multipleOf with a float', () => {
         let errors: RJSFValidationError[];
         beforeAll(() => {
-          const result = validator.validateFormData({ price: 1.05 }, resolvedRootSchema);
+          const result = validator.validateFormData({ price: 1.05 }, rootSchema);
           errors = result.errors;
         });
         it('should not return an error', () => {
@@ -174,7 +195,7 @@ describe('AJV8PrecompiledValidator', () => {
         let errors: RJSFValidationError[];
         let errorSchema: ErrorSchema;
         beforeAll(() => {
-          const result = validator.validateFormData({ price: 0.14 }, resolvedRootSchema);
+          const result = validator.validateFormData({ price: 0.14 }, rootSchema);
           errors = result.errors;
           errorSchema = result.errorSchema;
         });
@@ -198,7 +219,7 @@ describe('AJV8PrecompiledValidator', () => {
         describe('formData is provided at top level', () => {
           beforeAll(() => {
             const formData = { passwords: { pass1: 'a', pass2: 'b' } };
-            const result = validator.validateFormData(formData, resolvedRootSchema);
+            const result = validator.validateFormData(formData, rootSchema);
             errors = result.errors;
             errorSchema = result.errorSchema;
           });
@@ -209,7 +230,7 @@ describe('AJV8PrecompiledValidator', () => {
         describe('formData is not provided at top level', () => {
           beforeAll(() => {
             const formData = { passwords: { pass1: 'a' } };
-            const result = validator.validateFormData(formData, resolvedRootSchema);
+            const result = validator.validateFormData(formData, rootSchema);
             errors = result.errors;
             errorSchema = result.errorSchema;
           });
@@ -227,7 +248,7 @@ describe('AJV8PrecompiledValidator', () => {
         let errors: RJSFValidationError[];
 
         beforeAll(() => {
-          const result = validator.validateFormData({ anything: { foo: '42' } }, resolvedRootSchema);
+          const result = validator.validateFormData({ anything: { foo: '42' } }, rootSchema);
           errors = result.errors;
         });
 
@@ -240,7 +261,7 @@ describe('AJV8PrecompiledValidator', () => {
         let errorSchema: ErrorSchema;
 
         beforeAll(() => {
-          const result = validator.validateFormData({ anything: { foo: 42 } }, resolvedRootSchema);
+          const result = validator.validateFormData({ anything: { foo: 42 } }, rootSchema);
           errors = result.errors;
           errorSchema = result.errorSchema;
         });
@@ -267,13 +288,7 @@ describe('AJV8PrecompiledValidator', () => {
           transformErrors = jest.fn((errors: RJSFValidationError[]) => {
             return [Object.assign({}, errors[0], { message: newErrorMessage })];
           });
-          const result = validator.validateFormData(
-            { name: 42 },
-            resolvedRootSchema,
-            undefined,
-            transformErrors,
-            uiSchema
-          );
+          const result = validator.validateFormData({ name: 42 }, rootSchema, undefined, transformErrors, uiSchema);
           errors = result.errors;
         });
 
@@ -305,7 +320,7 @@ describe('AJV8PrecompiledValidator', () => {
         describe('formData is provided and passes custom validation', () => {
           beforeAll(() => {
             const formData = { passwords: { pass1: 'a', pass2: 'a' } };
-            const result = validator.validateFormData(formData, resolvedRootSchema, validate, undefined, uiSchema);
+            const result = validator.validateFormData(formData, rootSchema, validate, undefined, uiSchema);
             errors = result.errors;
             errorSchema = result.errorSchema;
           });
@@ -319,7 +334,7 @@ describe('AJV8PrecompiledValidator', () => {
         describe('formData is provided, but fails custom validation', () => {
           beforeAll(() => {
             const formData = { passwords: { pass1: 'a', pass2: 'b' } };
-            const result = validator.validateFormData(formData, resolvedRootSchema, validate, undefined, uiSchema);
+            const result = validator.validateFormData(formData, rootSchema, validate, undefined, uiSchema);
             errors = result.errors;
             errorSchema = result.errorSchema;
           });
@@ -338,7 +353,7 @@ describe('AJV8PrecompiledValidator', () => {
         describe('formData is missing data', () => {
           beforeAll(() => {
             const formData = { passwords: { pass1: 'a' } };
-            const result = validator.validateFormData(formData, resolvedRootSchema, validate);
+            const result = validator.validateFormData(formData, rootSchema, validate);
             errors = result.errors;
             errorSchema = result.errorSchema;
           });
@@ -359,21 +374,21 @@ describe('AJV8PrecompiledValidator', () => {
           const formData = {
             dataUrlWithName: 'data:text/plain;name=file1.txt;base64,x=',
           };
-          const result = validator.validateFormData(formData, resolvedRootSchema);
+          const result = validator.validateFormData(formData, rootSchema);
           expect(result.errors).toHaveLength(0);
         });
         it('Data-Url without name is accepted', () => {
           const formData = {
             dataUrlWithName: 'data:text/plain;base64,x=',
           };
-          const result = validator.validateFormData(formData, resolvedRootSchema);
+          const result = validator.validateFormData(formData, rootSchema);
           expect(result.errors).toHaveLength(0);
         });
         it('Data-Url with bad data generates error', () => {
           const formData = {
             dataUrlWithName: 'x=',
           };
-          const result = validator.validateFormData(formData, resolvedRootSchema);
+          const result = validator.validateFormData(formData, rootSchema);
           expect(result.errors).toHaveLength(1);
           expect(result.errorSchema.dataUrlWithName!.__errors).toHaveLength(1);
           expect(result.errorSchema.dataUrlWithName!.__errors![0]).toEqual('must match format "data-url"');
@@ -383,19 +398,16 @@ describe('AJV8PrecompiledValidator', () => {
   });
   describe('validator.validateFormData(), custom options, and localizer', () => {
     let validator: AJV8PrecompiledValidator;
-    let resolvedRootSchema: RJSFSchema;
     let localizer: Localizer;
     beforeAll(() => {
       localizer = jest.fn().mockImplementation();
       validator = new AJV8PrecompiledValidator(validateOptionsFns, rootSchema, localizer);
-      // Since the root schema is retrieved in core before calling the validator emulate that here
-      resolvedRootSchema = retrieveSchema(validator, rootSchema, rootSchema);
     });
     describe('validating using single custom meta schema', () => {
       let errors: RJSFValidationError[];
       beforeAll(() => {
         (localizer as jest.Mock).mockClear();
-        const result = validator.validateFormData({ foo: 42 }, resolvedRootSchema);
+        const result = validator.validateFormData({ foo: 42 }, rootSchema);
         errors = result.errors;
       });
       it('should return 1 error about formData', () => {
@@ -425,14 +437,14 @@ describe('AJV8PrecompiledValidator', () => {
     });
     describe('validating using custom string formats', () => {
       it('should not return a validation error if proper string format is used', () => {
-        const result = validator.validateFormData({ phone: '800-555-2368' }, resolvedRootSchema);
+        const result = validator.validateFormData({ phone: '800-555-2368' }, rootSchema);
         expect(result.errors).toHaveLength(0);
       });
       describe('validating using a custom formats', () => {
         let errors: RJSFValidationError[];
 
         beforeAll(() => {
-          const result = validator.validateFormData({ phone: '800.555.2368' }, resolvedRootSchema);
+          const result = validator.validateFormData({ phone: '800.555.2368' }, rootSchema);
           errors = result.errors;
         });
         it('should return 1 error about formData', () => {